Move prepareAutoUploadImages to utils
